% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIMPLE-package.R
\docType{package}
\name{CIMPLE-package}
\alias{CIMPLE}
\alias{CIMPLE-package}
\title{CIMPLE: CIMPLE: Clinically Informative Missingness through Probabilities, Likelihoods, and Estimating Equations}
\description{
This package offers a collection of methods involved with analyzing longitudinal EHR data with possibly informative observational time. These methods are grouped into two classes depending on the inferential task. One group focuses on estimating the effect of an exposure on a longitudinal biomarker while the other group assesses the impact of a longitudinal biomarker on time-to-diagnosis outcomes.
}
\author{
\strong{Maintainer}: Jiacong Du \email{jiacong@umich.edu} (\href{https://orcid.org/0000-0002-7768-9083}{ORCID})

Other contributors:
\itemize{
  \item Howard Baik \email{howard.baik@yale.edu} (\href{https://orcid.org/0009-0000-8942-1618}{ORCID}) [contributor]
}

}
\keyword{internal}
