% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_est.R
\name{long_est}
\alias{long_est}
\title{Coefficient estimation in the longitudinal model}
\usage{
long_est(
  long_data,
  method,
  ...,
  id_var,
  outcome_var,
  LM_fixedEffect_variables = NULL,
  time = NULL,
  LM_randomEffect_variables = NULL,
  VPM_variables = NULL,
  imp_time_factor = NULL,
  optCtrl = list(method = "nlminb", kkt = FALSE, tol = 0.2, maxit = 20000),
  control = list(verbose = FALSE, tol = 0.001, GHk = 10, maxiter = 150)
)
}
\arguments{
\item{long_data}{Long dataset}

\item{method}{The following methods are available:
\itemize{
\item \code{standard_LME}: Standard linear mixed-effect model.
\item \code{VA_LME}: Linear mixed-effect model adjusted for the historical number of visits.
\item \code{JMVL_LY}: Joint model of the visiting process and the longitudinal process accounting for measured confounders.
\item \code{IIRR_weighting}: Inverse-intensity-rate-ratio weighting approach.
\item \code{JMVL_Liang}: Joint model of the visiting process and the longitudinal process with dependent latent variables.
\item \code{imputation_LME}: Imputation-based approach with linear mixed-effect model.
\item \code{JMVL_G}: Joint model of the visiting process and the longitudinal process with a shared random intercept.
}}

\item{...}{Additional arguments to \code{\link[nleqslv:nleqslv]{nleqslv::nleqslv()}}.}

\item{id_var:}{Variable for the subject ID to indicate the grouping
structure.}

\item{time:}{Variable for the observational time should be named as 'time'.}

\item{outcome_var:}{Variable name for the longitudinal outcome variable.}

\item{LM_fixedEffect_variables:}{Vector input of variable names with fixed
effects in the longitudinal model. Variables should not contain \code{time}
(Note: is the correct?). TODO: Note: above}

\item{LM_randomEffect_variables:}{Vector input of variable names with random
effects in the longitudinal model. This argument is NULL for methods
including \code{JMVL_LY}, \code{JMVL_G} and \code{IIRR_weighting}.}

\item{VPM_variables:}{Vector input of variable names in the visiting process
model.}

\item{imp_time_factor:}{scale factor for the time variable. This argument is
only needed in the imputation-based methods.}

\item{optCtrl:}{control parameters for running the mixed-effect model. See
the \code{control} argument in \code{\link[lme4:lmer]{lme4::lmer()}}.}

\item{control:}{control parameters for the \code{JMVL_G} method, including: (1)
\code{verbose}: TRUE or FALSE for outputting checkpoint after each iteration.
Default is FALSE. (2) \code{tol}: tolerance for convergence (3) \code{GHk}: number of
gaussian-hermite quadrature points. Default is 10. (5) \code{maxiter}: maximum
number of iteration. Default is 150.}
}
\value{
\code{beta_hat}: Estimated coefficients in the longitudinal model.

Other output in each method:
\itemize{
\item \code{standard_LME}:
\itemize{
\item \code{beta_sd}: Standard deviation of the estimated coefficients.
}
\item \code{VA_LME}:
\itemize{
\item \code{beta_sd}: Standard deviation of the estimated coefficients.
}
\item \code{JMVL_LY}:
\itemize{
\item \code{gamma_hat}: Estimated coefficients in the visiting process model.
}
\item \code{IIRR_weighting}:
\itemize{
\item \code{gamma_hat}: Estimated coefficients in the visiting process model.
}
\item \code{JMVL_Liang}:
\itemize{
\item \code{gamma_hat}: Estimated coefficients in the visiting process model.
}
}
}
\description{
This function offers a collection of methods of coefficient estimation in a
longitudinal model with possibly informative observation time.
These methods include Standard linear mixed-effect model (Standard_LME),
Linear mixed-effect model adjusted for the historical number of visits (VA_LME),
Joint model of the visiting process and the longitudinal process accounting for measured confounders (JMVL_LY),
Inverse-intensity-rate-ratio weighting approach (IIRR_weighting),
Joint model of the visiting process and the longitudinal process with dependent latent variables (JMVL_Liang),
Imputation-based approach with linear mixed-effect model (imputation_LME), and
Joint model of the visiting process and the longitudinal process with a shared random intercept (JMVL_G).
}
\examples{
# Setup arguments
train_data

time_var = "time"
id_var = "id"
outcome_var = "Y"
VPM_variables = c("Z", "X")
LM_fixedEffect_variables = c("Z", "X")
LM_randomEffect_variables = "Z"

# Run the standard LME model
fit_standardLME = long_est(long_data=train_data,
                           method="standard_LME",
                           id_var=id_var,
                           outcome_var=outcome_var,
                           LM_fixedEffect_variables = LM_fixedEffect_variables,
                           time = time_var,
                           LM_randomEffect_variables = LM_randomEffect_variables,
                           VPM_variables = VPM_variables)
# Return the coefficient estimates
fit_standardLME$beta_hat

# Run the VA_LME model
fit_VALME = long_est(long_data=train_data,
                     method="VA_LME",
                     id_var=id_var,
                     outcome_var=outcome_var,
                     LM_fixedEffect_variables = LM_fixedEffect_variables,
                     time = time_var,
                     LM_randomEffect_variables = LM_randomEffect_variables,
                     VPM_variables = VPM_variables)
# Return the coefficient estimates
fit_VALME$beta_hat
}
